pyproject.toml
[project]
name = "speculative-beam-decoding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "torch>=2.7.0",
    "transformers>=4.52.2",
    "wandb>=0.19.11",
]
